Slight Changes
Removed Bstats integration, I thought it would be cool to be able to see how popular my plugin is, but i rarely check it, so to make it easier on server owners, i just decided to remove it entirely.
Corrected Version manager, as it checks float version rather than sub version. So if  a server was on 1.9, it would state that it could support custom colors when it cannot. because 1.9 > 1.16, massive bug
Spigot issue that i never caught that will complain for some reason about null ItemMeta, forgot to add a check in place, still works and no error from my testing. I use papermc for my server for testing, and this issue has only occured on spigot, so idk
Update Manager still is in the code, but will not be called as the current version is deprecated by Spigot.

I might have to make a new version being 1.7.2 so that these changes are implemented into Aternos